@@ -1,14 +1,14 @@
 package edu.utrack.activity.dataview;
 
-import android.support.v4.app.Fragment;
 
-import java.util.List;
-import java.util.Map;
-
+import androidx.fragment.app.Fragment;
 import edu.utrack.data.app.AppEvent;
 import edu.utrack.data.calendar.CalendarEvent;
 import edu.utrack.data.screen.ScreenEvent;
 import edu.utrack.data.screen.ScreenEventType;
+
+import java.util.List;
+import java.util.Map;
 
 /**
  * Created by Tobi on 29/03/2018.
